samples-ImageProcessing/without-ml/human-faces-dataset-creator/src/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

addict
future
lmdb
numpy
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
tqdm
yapf
torch
torchvision
insightface
onnxruntime-gpu
